Add oil filter notes
Looks like there are domestic options available for oil filters.

There's no real reason to not use OEM as the price is good and the quality
should be acceptable, but if one is needed in a pinch or if local stock
of OEM becomes hard to get or expensive then this is good to know.